+++ doc/release-notes/v9/9.1.0/04_divers.sql
@@ -1,0 +1,9 @@
+--
+-- 9.1.0
+--
+
+--
+-- Contrainte de référence ajoutée récemment à tort (empêche la suppression d'un manuscrit déposé).
+--
+
+alter table validite_fichier drop constraint validite_fichier_ffk;